pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nicechain"
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Kurtis Fields", email = "kurtisfields@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "loguru ==0.7.0",
  "python-dotenv ==1.0.0",
  "click ==8.1.3",
  "nicegui ==1.2.13",
  "openai ==0.27.6",
  "cohere ==4.3.1",
  "google-generativeai ==0.1.0rc2",
  "langchain ==0.0.162",
  "PyMuPDF ==1.22.2",
  #"redis ==4.5.4",
  "chromadb ==0.3.22",
  "tiktoken ==0.4.0",
  "lark ==1.1.5"
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/kfields/nicechain#readme"
Issues = "https://github.com/kfields/nicechain/issues"
Source = "https://github.com/kfields/nicechain"

[tool.hatch.version]
path = "nicechain/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=nicechain --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "nicechain/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
